+++ bin/handlers/get_switches
@@ -4,7 +4,7 @@ Command line handler to get the number of started switches
 from the distributed topologies
 """
 
-import util.multinet_util as m_util
+import util.multinet_requests as m_util
 
 
 def get_switches_main():
