+++ bin/handlers/pingall
@@ -4,7 +4,7 @@ Command line handler to do a pingall in each one
 of the distributed topologies 
 """
 
-import util.multinet_util as m_util
+import util.multinet_requests as m_util
 
 
 def pingall_handler_main():
