+++ bin/handlers/start_topos
@@ -3,7 +3,7 @@
 Command line handler to start the distributed topologies
 """
 
-import util.multinet_util as m_util
+import util.multinet_requests as m_util
 
 
 def start_handler_main():
